\name{compcrab}
\alias{compcrab}
\docType{data}
\title{
  Catch comparison on unpaired hauls 
}
\description{
  Data for catch comparison on unpaired hauls for the unwanted bycatch edible crab in plaice gillnet in \emph{compcrab}
}
\usage{data("compcrab")}
\format{
  A data frame with 563 observations on the following 4 variables.
  \describe{
    \item{\code{width}}{width: a numeric for carapace width}
    \item{\code{haul}}{haul: a factor with the form \code{Day_Soak_Fleet}
      representing each deployment of a fleet, i.e., a haul}
    \item{\code{tacti}}{tactic: a factor with 2 levels for soak tactic (day or night)}
    \item{\code{total}}{total: a numeric for the total number of individuals for each length class and haul}
  }
}
\source{
  From Esther Savina
}
\references{
  Savina, E., Krag, L.A., Frandsen, R.P., and Madsen, N. (2017). Effect of fisher's soak tactic on catch pattern in the Danish gillnet plaice fishery.
  \emph{Fisehries Research} \bold{196}, 56--65;
  \url{https://doi.org/10.1016/j.fishres.2017.08.009}
}

\keyword{datasets}
