% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expandAllGrpVar}
\alias{expandAllGrpVar}
\title{expand interactions/combinations of grouping variables}
\usage{
expandAllGrpVar(bb)
}
\arguments{
\item{bb}{a list of naked grouping variables, i.e. 1 | f}
}
\description{
Modeled after lme4:::expandSlash, by Doug Bates
}
\keyword{internal}
