% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getME.selfisher}
\alias{getME.selfisher}
\alias{getME}
\title{Extract or Get Generalize Components from a Fitted Mixed Effects Model}
\usage{
\method{getME}{selfisher}(
  object,
  name = c("Xr", "Xp", "Zr", "Zp", "Xd", "betar", "betap", "betad", "br", "bp",
    "thetar", "thetap"),
  ...
)
}
\arguments{
\item{object}{a fitted \code{selfisher} object}

\item{name}{of the component to be retrieved}

\item{\dots}{ignored, for method compatibility}
}
\description{
Extract or Get Generalize Components from a Fitted Mixed Effects Model
}
\seealso{
\code{\link[lme4]{getME}}
Get generic and re-export:
}
