% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{inits}
\alias{inits}
\title{Get initial values for intercept (a) and coefficient on the length term (b)
based on the link funciton, and your guesses for L50 and SR}
\usage{
inits(L50, SR, link = "logit", d = 1)
}
\arguments{
\item{L50}{guess for what L50 might be}

\item{SR}{guess for what SR might be}

\item{link}{character}

\item{d}{delta exponent in Richards link model}
}
\description{
Get initial values for intercept (a) and coefficient on the length term (b)
based on the link funciton, and your guesses for L50 and SR
}
