% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootSel.R
\name{L100}
\alias{L100}
\title{a function taking a fitted \code{selfisher} object as input and returning the
L1 to 100 (i.e. the lengths with catch probabilities 0.01 to 1.0)}
\usage{
L100(x)
}
\arguments{
\item{x}{\code{selfisher} object fitted with arugument code{Lp="100"}}
}
\description{
a function taking a fitted \code{selfisher} object as input and returning the
L1 to 100 (i.e. the lengths with catch probabilities 0.01 to 1.0)
}

