% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfisher.R
\name{mkTMBStruc}
\alias{mkTMBStruc}
\title{Extract info from formulas, reTrms, etc., format for TMB}
\usage{
mkTMBStruc(rformula, pformula, dformula, combForm, mf, fr, yobs, total, family,
  link_char, cover, Lp, pPredictCode = "selection", doPredict = 0,
  whichPredict = integer(0), x0 = NULL, call = NULL, respCol)
}
\arguments{
\item{rformula}{selectivity formula}

\item{pformula}{relative fishing power formula}

\item{dformula}{Richards delta parameter formula}

\item{combForm}{combined formula}

\item{mf}{call to model frame}

\item{fr}{frame}

\item{yobs}{observed y}

\item{total}{total}

\item{cover}{(logical) covered codend model (i.e. big fish go in experimental net and small fish go in covered codend)}

\item{Lp}{controls calculation of length (L) at retention prob (p)}

\item{pPredictCode}{relative fishing power code}

\item{doPredict}{flag to enable sds of predictions}

\item{whichPredict}{which observations in model frame represent predictions}

\item{x0}{vector of initial values for the size selectivity model}

\item{respCol}{response column}

\item{link}{character}
}
\description{
Extract info from formulas, reTrms, etc., format for TMB
}
\keyword{internal}

