% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.selfisher}
\alias{predict.selfisher}
\title{prediction}
\usage{
\method{predict}{selfisher}(object, newdata = NULL, se.fit = FALSE, re.form,
  allow.new.levels = FALSE, type = c("response", "selection", "prob"),
  debug = FALSE, ...)
}
\arguments{
\item{object}{a \code{selfisher} object}

\item{newdata}{new data for prediction}

\item{se.fit}{return the standard errors of the predicted values?}

\item{re.form}{(not yet implemented) specify which random effects to condition on when predicting}

\item{allow.new.levels}{(not yet implemented) allow previously unobserved levels in random-effects grouping variables?}

\item{type}{return expected value ("response": pr/(pr+1-p)),
the  predicted selection curve ("selection": r),
or the relative fishing power of the test gear ("prob")?}

\item{debug}{(logical) return the \code{TMBStruc} object that will be
used internally for debugging?}

\item{\dots}{unused - for method compatibility}
}
\description{
prediction
}

