% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarCorr.R
\name{print.VarCorr.selfisher}
\alias{print.VarCorr.selfisher}
\title{Printing The Variance and Correlation Parameters of a \code{selfisher}}
\usage{
\method{print}{VarCorr.selfisher}(
  x,
  digits = max(3, getOption("digits") - 2),
  comp = "Std.Dev.",
  formatter = format,
  ...
)
}
\arguments{
\item{x}{a result of \code{\link{VarCorr}(<selfisher>)}.}

\item{digits}{number of significant digits to use.}

\item{comp}{a string specifying the component to format and print.}

\item{formatter}{a \code{\link{function}}.}

\item{...}{optional further arguments, passed the next \code{\link{print}} method.}
}
\description{
Printing The Variance and Correlation Parameters of a \code{selfisher}
}
