% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.selfisher}
\alias{residuals.selfisher}
\title{Compute residuals for a selfisher object}
\usage{
\method{residuals}{selfisher}(object, type = c("response", "pearson",
  "deviance"), ...)
}
\arguments{
\item{object}{a \dQuote{selfisher} object}

\item{type}{(character) residual type}

\item{\dots}{ignored, for method compatibility}
}
\description{
Compute residuals for a selfisher object
}
