% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarCorr.R
\name{richardsdelta}
\alias{richardsdelta}
\title{Extract Richards exponent parameter}
\usage{
richardsdelta(object, ...)
}
\arguments{
\item{object}{a \dQuote{selfisher} fitted object}

\item{\dots}{(ignored; for method compatibility)}
}
\description{
Extract Richards exponent parameter
}

