% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfisher.R
\name{selfisher}
\alias{selfisher}
\title{Fit gear selectivity models with TMB}
\usage{
selfisher(
  rformula,
  data = NULL,
  pformula = ~1,
  dformula = ~1,
  psplit = FALSE,
  start = NULL,
  link = "logit",
  total = NULL,
  haul = NULL,
  pool = NULL,
  qratio = NULL,
  offset = NULL,
  Lp = "basic",
  se = TRUE,
  verbose = FALSE,
  debug = FALSE,
  optControl = list(iter.max = 300, eval.max = 400)
)
}
\arguments{
\item{rformula}{combined fixed and random effects formula for the selectivity model, following lme4
syntax. The left-hand side of the formula should be the proportion of fish retained in the test gear.}

\item{data}{data frame}

\item{pformula}{a \emph{one-sided} (i.e., no response variable) formula for
    the ralative fishing power of the test versus the control gear combining fixed and random effects:
\code{~0} can be used to specify equal fishing power (p=0.5).
\code{~1} estimates one overall relative fishing power.
\code{~(1|haul)} allows relative fishing power to vary randomly by \code{haul}.
The relative fishing power model uses a logit link.}

\item{dformula}{a formula for the delta parameter in Richards selection curve. Ignored unless \code{link="richards"}.}

\item{psplit}{(logical) Use TRUE for paired gear experiments where one gear is non-selective.}

\item{start}{starting values, expressed as a list with possible components
\code{betar}, \code{betap}, \code{betad} (fixed-effect parameters for
retention, psplit, Richards delta models); \code{br}, \code{bp}
(conditional modes for retention and psplit models);
\code{thetar}, \code{thetap} (random-effect parameters, on the
standard deviation/Cholesky scale, for retention and psplit models);}

\item{link}{A character indicating the link function for the selectivity model.
\code{"logit"}(logistic) is the default, but other options are "probit" (i.e. normal probability ogiv), "cloglog" (i.e. negative extreme value), "loglog" (i.e. extreme value/Gompert), or "Richards"}

\item{total}{Name of column representing the number of total fish caught and counted in one length class within haul, i.e. the binomial denominator (e.g. codend + cover, test + control, or test1 + test2).}

\item{haul}{Name of column representing different hauls. Needed for double bootstrap methods.}

\item{pool}{(Optional) name of column representing different pools of hauls. Used in hierarchical bootstrapping to produce same number of hauls by pool.}

\item{qratio}{Name of column representing the ratio of sampling fractions, e.g. (sampled test/sampled control) or (sampled codend/sampled cover) or in the general binomial definition (sampled successes/sampled failures). Use this to correct for sampling.}

\item{offset}{see details below}

\item{Lp}{controls calculation of length (l) at retention prob (p), see details}

\item{se}{(logical) whether to return standard errors}

\item{verbose}{(logical) progress indication should be printed to the console.}

\item{debug}{(logical) whether to return the preprocessed data and parameter objects,
without fitting the model}

\item{optControl}{control parameters passed to \code{nlminb}}
}
\description{
Fit gear selectivity models with TMB
}
\details{
\itemize{
\item In general, \code{qratio} is more applicable than traditional GLM offsets. Do not use an offset together with \code{qratio}. Always use \code{qratio} instead of an offset when using a link other than logit, or when \code{psplit=TRUE}.
\item In all cases \code{selfisher} returns maximum likelihood estimates.
\item Lp="basic" will return values for l50 and SR.
\item Lp="none" supresses calculation of l50 and SR to save time.
\item Use \code{getCapabilities()} to see options for retention model links and random effects covariance structures.
}
}
\examples{
dat <- transform(haddock, tot=nfine+nwide, prop=nwide/(nfine+nwide))
m0 <- selfisher(prop~Lengths, pformula=~0, psplit=TRUE, total=tot, dat)
m1 <- selfisher(prop~Lengths, pformula=~1, psplit=TRUE, total=tot, dat)
}
\references{
More examples are available in Brooks ME, Melli V, Savina E, Santos J, Millar R, O'Neill FG, Veiga-Malta T, Krag LA, Feekings JP. 2022. Introducing selfisher: open source software for statistical analyses of fishing gear selectivity. Canadian Journal of Fisheries and Aquatic Sciences. https://doi.org/10.1139/cjfas-2021-0099
}
