% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarCorr.R
\name{VarCorr.selfisher}
\alias{VarCorr}
\alias{VarCorr.selfisher}
\title{Extract variance and correlation components}
\usage{
\method{VarCorr}{selfisher}(x, sigma = 1, ...)
}
\arguments{
\item{x}{a fitted \code{selfisher} model}

\item{sigma}{residual standard deviation (usually set automatically from internal information)}

\item{extra}{arguments (for consistency with generic method)}
}
\description{
Extract variance and correlation components
}
\examples{
## Comparing variance-covariance matrix with manual computation
}
\keyword{internal}

