% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{vcov.selfisher}
\alias{vcov.selfisher}
\title{Calculate Variance-Covariance Matrix for a Fitted selfisher model}
\usage{
\method{vcov}{selfisher}(object, full = FALSE, ...)
}
\arguments{
\item{object}{a \dQuote{selfisher} fit}

\item{full}{return a full variance-covariance matrix?}

\item{\dots}{ignored, for method compatibility}
}
\value{
By default (\code{full==FALSE}), a list of separate variance-covariance matrices for each model component (conditional, zero-inflation, dispersion).  If \code{full==TRUE}, a single square variance-covariance matrix for \emph{all} top-level model parameters (conditional, dispersion, and variance-covariance parameters)
}
\description{
Calculate Variance-Covariance Matrix for a Fitted selfisher model
}

