% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootSel.R
\name{bootSel}
\alias{bootSel}
\title{Perform bootstrap}
\usage{
bootSel(x, FUN = L50SR, nsim = 2, seed = NULL, type = c("double",
  "parametric", "nonparameteric"), verbose = FALSE, .progress = "none",
  PBargs = list(), parallel = c("no", "multicore", "snow"),
  ncpus = getOption("boot.ncpus", 1L), cl = NULL)
}
\arguments{
\item{x}{a fitted \code{selfisher} object}

\item{FUN}{a function taking a fitted
\code{selfisher} object as input and returning the
\emph{statistic} of interest, which must be a (possibly named) numeric vector.}

\item{nsim}{number of simulations, positive integer}

\item{seed}{optional argument to \code{\link{set.seed}}}

\item{type}{character string specifying the type of
bootstrap, \code{double}(the defualt) as defined in gear selectivity literature (e.g. Millar 1993),
\code{"parametric"} or \code{"nonparametric"}; partial matching is allowed.}
}
\description{
Perform bootstrap
}
\details{
the default bootstrap type "double" is specific to fisheries literature.
}

