% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootSel.R
\name{bootSel}
\alias{bootSel}
\title{Perform bootstrap}
\usage{
bootSel(
  x,
  FUN = L50SR,
  nsim = 2,
  seed = NULL,
  type = c("double", "parametric", "nonparameteric"),
  verbose = FALSE,
  .progress = "none",
  PBargs = list(),
  parallel = c("no", "multicore", "snow"),
  ncpus = getOption("boot.ncpus", 1L),
  cl = NULL
)
}
\arguments{
\item{x}{a fitted \code{selfisher} object}

\item{FUN}{a function taking a fitted
\code{selfisher} object as input and returning the
\emph{statistic} of interest, which must be a (possibly named) numeric vector.
The default (\code{FUN = L50SR}) only works with very simple models; use \code{\link{predict.selfisher}}.}

\item{nsim}{number of simulations, positive integer}

\item{seed}{optional argument to \code{\link{set.seed}}}

\item{type}{character string specifying the type of
bootstrap, \code{double}(the defualt) as defined in gear selectivity literature (Millar 1993),
\code{"parametric"} or \code{"nonparametric"}; partial matching is allowed. Only the default version has been tested.}
}
\description{
Perform bootstrap
}
\details{
The code is based on code from the lme4 package,
except that the default bootstrap type "double"
is specific to fisheries literature.
This code has not been tested on models containing random effects;
we recommend removing them before bootstrapping. The double bootstrap
procedure accounts for variability among "hauls" and it should be possible to
use this to account for any factor that could be treated as a random effect.
See \code{vignette("bootstrap")} for an example.
}
