% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expandAllGrpVar}
\alias{expandAllGrpVar}
\title{expand interactions/combinations of grouping variables}
\usage{
expandAllGrpVar(bb)
}
\arguments{
\item{bb}{a list of naked grouping variables, i.e. 1 | f}
}
\description{
Modeled after lme4:::expandSlash, by Doug Bates
}
\examples{
ff <- selfisher:::fbx(y~1+(x|f/g))
selfisher:::expandAllGrpVar(ff)
selfisher:::expandAllGrpVar(quote(1|(f/g)/h))
selfisher:::expandAllGrpVar(quote(1|f/g/h))
selfisher:::expandAllGrpVar(quote(1|f*g))
}
\keyword{internal}

