% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{inits}
\alias{inits}
\title{Get initial values for intercept (a) and coefficient on the length term (b)
based on the link funciton, and your guesses for L50 and SR}
\usage{
inits(L50, SR, link = "logit", d = 1)
}
\arguments{
\item{L50}{guess for what L50 might be}

\item{SR}{guess for what SR might be}

\item{link}{character}

\item{d}{delta exponent in Richards link model}
}
\value{
c(a,b) where a = intercept and b = slope of length
}
\description{
Get initial values for intercept (a) and coefficient on the length term (b)
based on the link funciton, and your guesses for L50 and SR
}
\examples{
dat = transform(haddock, tot=nfine+nwide, prop=nwide/(nfine+nwide))
m1 = selfisher(prop~Lengths, pformula=~1, psplit=TRUE, total=tot, dat, 
  link="cloglog", start=list(betar=inits(29.428195,  1.780573, link="cloglog")))
}
