% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfisher.R
\name{mkTMBStruc}
\alias{mkTMBStruc}
\title{Extract info from formulas, reTrms, etc., format for TMB}
\usage{
mkTMBStruc(rformula, pformula, dformula, mf, fr, yobs, offset, weights, family,
  link, pPredictCode = "selection", doPredict = 0,
  whichPredict = integer(0))
}
\arguments{
\item{rformula}{selectivity formula}

\item{pformula}{relative fishing power formula}

\item{dformula}{Richards delta parameter formula}

\item{mf}{call to model frame}

\item{fr}{frame}

\item{yobs}{observed y}

\item{offset}{offset}

\item{weights}{weights}

\item{link}{character}

\item{pPredictCode}{relative fishing power code}

\item{doPredict}{flag to enable sds of predictions}

\item{whichPredict}{which observations in model frame represent predictions}
}
\description{
Extract info from formulas, reTrms, etc., format for TMB
}
\keyword{internal}

