% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfisher.R
\name{selfisher}
\alias{selfisher}
\title{Fit models with TMB}
\usage{
selfisher(rformula, pformula = ~0, dformula = ~1, data = NULL,
  link = "logit", weights = NULL, offset = NULL, se = TRUE,
  verbose = FALSE, debug = FALSE)
}
\arguments{
\item{rformula}{combined fixed and random effects formula for the selectivity model, following lme4
syntax. The left-hand side of the formula should be the proportion of fish entering the test gear.}

\item{pformula}{a \emph{one-sided} (i.e., no response variable) formula for
    the ralaive fishing power of the test versus the control gear combining fixed and random effects:
the default \code{~0} specifies equal fishing power (p=0.5).
The relative fishing power model uses a logit link.}

\item{dformula}{a formula for the delta parameter in Richards selection curve. Ignored unless \code{link="richards"}.}

\item{data}{data frame}

\item{link}{A character indicating the link function for the selectivity model. 
\code{"logit"} is the default, but other options can be used (use \code{getCapabilities()} to see options).}

\item{weights}{The number of total fish caught in the test and control gear.}

\item{offset}{offset}

\item{se}{whether to return standard errors}

\item{verbose}{logical indicating if some progress indication should be printed to the console.}

\item{debug}{whether to return the preprocessed data and parameter objects,
without fitting the model}
}
\description{
Fit models with TMB
}
\details{
\itemize{
\item in all cases \code{selfisher} returns maximum likelihood estimates - random effects variance-covariance matrices are not REML (so use \code{REML=FALSE} when comparing with \code{lme4::lmer}), and residual standard deviations (\code{\link{sigma}}) are not bias-corrected. Because the \code{\link{df.residual}} method for \code{selfisher} currently counts the dispersion parameter, one would need to multiply by \code{sqrt(nobs(fit)/(1+df.residual(fit)))} when comparing with \code{lm} ...
}
}

